chore: move the app to the api folder
