+++ .github/workflows/node_ci.yml
@@ -11,7 +11,7 @@ jobs:
 
     strategy:
       matrix:
-        node-version: [18]
+        node-version: [18.x]
 
     steps:
       - uses: actions/checkout@v3
