+++ src/dev-server.ts
@@ -1,0 +1,5 @@
+import app from '../api'
+
+app.listen(8000, () => {
+  console.log('http://localhost:8000')
+})
