+++ test/invalid-routes.test.ts
@@ -1,5 +1,5 @@
 import request from 'supertest'
-import app from '../src/app'
+import app from '../api'
 
 describe('Missing username query', () => {
   it('should return 400 status with info message', async () => {
